api/app/pyproject.toml
[tool.poetry]
name = "Clients api"
version = "1"
description = ""
authors = ["Kuzin Danil"]


[tool.poetry.dependencies]
python = "^3.10"
python-multipart = "^0.0.10"
fastapi = "^0.113"
uvicorn = "^0.30"
SQLAlchemy = "^2.0.23"
asyncpg = "^0.29.0"
pydantic = "^2.5.2"
pydantic-settings = "^2.1.0"
pylint = "^3.0.2"
bcrypt = "^4.0.0"
PyJWT = "^2.8.0"
pillow = "^10.0.0"

[build-system]
requires = ["poetry-core>=1.8.1"]
build-backend = "poetry.core.masonry.api"
